pyproject.toml
[project]
name = "code-rosetta-mcp"
version = "0.1.0"
description = "Cross-language codebase graph for LLM agents. Python, Terraform, YAML, Jinja2 — one unified map."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "tree-sitter>=0.23.0",
    "tree-sitter-language-pack>=0.10.0",
    "python-hcl2>=6.0.0",
    "ruamel.yaml>=0.18.0",
    "jinja2>=3.1.0",
    "networkx>=3.0",
    "fastmcp>=1.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[project.scripts]
code-rosetta = "code_rosetta.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["code_rosetta"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
